Add Functions to Separate module and resolve using global map
** Description **

Previously, when we defined a function, and when we called the
function. The function definition and anon_func for call were
both in the same module, when we executed the anon_call
Top level expression, we call the function one (executed)
then deleted the module. This delete both the anon func which
had the call, but then also the actual function defintion.

As a result We add each function defined to a separate module. When we call the
function, if the function is not defined in that module, we will use
the prototye stored in the global map to generate a declaration (not
definition)
JIT can find the module during cross module resolution.

// High Level
1. User Input: def foo(x) x + 1;
- Compile foo into Module 1.
- Add Module 1 to the JIT.

2. User Input: foo(2);

- Create an anonymous function calling foo(2) in Module 2.
- Regenerate foo's prototype in Module 2.
- Add Module 2 to the JIT.
- Execute the anonymous function.
- Remove Module 2 from the JIT.

// How does Cross-Module Function Calls happen?
The most interesting part is how test1 (which was defined earlier) gets called from the anonymous function. The JIT enables this through:

- Function Prototyping: When generating code for test1(4), the system looks up or regenerates the prototype for test1 in the current module.
- Cross-Module Resolution: The JIT resolves function calls across module boundaries by searching all previously loaded modules from newest to oldest.
- Dynamic Linking: When executing, the call to test1(4) is dynamically linked to the actual implementation of test1 in its own module.

** Testing **

ready> def foo(x) x + 1;
ready> Parsed a Function Definition.
define double @foo(double %x) {
entry:
  %addtmp = fadd double %x, 1.000000e+00
  ret double %addtmp
}

Adding Function to a separate module
ready> foo(2);
ready> Parsed a top-level expr
define double @__anon_expr() {
entry:
  %calltmp = call double @foo(double 2.000000e+00)
  ret double %calltmp
}

Evaluated to 3.000000
ready> foo(5);
ready> Parsed a top-level expr
define double @__anon_expr() {
entry:
  %calltmp = call double @foo(double 5.000000e+00)
  ret double %calltmp
}

Evaluated to 6.000000